MM-23: reorganize code for better readability
