+++ src/MAPIT.cpp
@@ -1,6 +1,12 @@
 // Copyright 2017-2021 Lorin Crawford.
 
 #include "MAPIT.h"
+#include "mapit/davies.h"
+#include "mapit/normal.h"
+#include "mapit/projection.h"
+#include "mapit/util.h"
+#include "gsm/gsm.h"
+#include "mqs/mqs.h"
 
 using std::chrono::steady_clock;
 using std::chrono::duration_cast;
@@ -9,112 +15,6 @@ using std::chrono::microseconds;
 // [[Rcpp::depends(RcppArmadillo)]]
 // [[Rcpp::plugins(cpp11)]]
 // [[Rcpp::plugins(openmp)]]
-
-arma::mat GetLinearKernel(arma::mat X) {
-    double p = X.n_rows;
-    return X.t() * X / p;
-}
-
-arma::mat ComputePCs(arma::mat X, int top = 10) {
-    arma::mat U;
-    arma::vec s;
-    arma::mat V;
-    svd(U, s, V, X);
-
-    arma::mat PCs = U*diagmat(s);
-    return PCs.cols(0, top - 1);
-}
-
-arma::mat ComputeProjectionMatrix(int n, arma::mat b) {
-    arma::mat identity = arma::eye<arma::mat>(n, n);
-    return identity - b * arma::inv(b.t() * b) * b.t();
-}
-
-
-// Computes Tr(ab)
-double ProductTrace(arma::mat a, arma::mat b) {
-    // Computed efficiently using Hadamard (elementwise) product
-    // https://en.wikipedia.org/wiki/Trace_(linear_algebra)#Trace_of_a_product
-    // https://proofwiki.org/wiki/Trace_of_Matrix_Product
-    return arma::as_scalar(accu(a%b));
-}
-
-arma::mat ComputeSMatrix(std::vector<arma::mat> matrices) {
-#ifdef WITH_LOGGER_FINE
-    std::string logname = "MAPITcpp::ComputeSMatrix";
-    auto logger = spdlog::get(logname);  // retrieve existing one
-     // or create new one if needed
-    if (logger == nullptr) logger = spdlog::r_sink_mt(logname);
-#endif
-    int num_variance_components = matrices.size();
-    arma::mat S = arma::zeros(num_variance_components, num_variance_components);
-
-    for (int i = 0; i < num_variance_components; i++) {
-        for (int j = 0; j < num_variance_components; j++) {
-            if (i <= j) {  // create upper triangular matrix
-                S(i, j) = ProductTrace(matrices[i], matrices[j]);
-                S(j, i) = S(i, j);
-#ifdef WITH_LOGGER_FINE
-                logger->info("S({},{}) = {}", i, j, S(i, j));
-                if (S(i, j) < 0) {
-                    logger->warn("S({},{}) negative", i, j);
-                }
-#endif
-            }
-        }
-    }
-    return S;
-}
-
-arma::vec ComputeqVector(arma::vec yc, std::vector<arma::mat> matrices) {
-#ifdef WITH_LOGGER_FINE
-    std::string logname = "MAPITcpp::ComputeqVector";
-    auto logger = spdlog::get(logname);
-    if (logger == nullptr) logger = spdlog::r_sink_mt(logname);
-#endif
-    int num_variance_components = matrices.size();
-    arma::vec q = arma::zeros(num_variance_components);
-
-    for (int i = 0; i < num_variance_components; i++) {
-        q(i) = arma::as_scalar(yc.t() * matrices[i] * yc);
-#ifdef WITH_LOGGER_FINE
-        logger->info("q({}) = {}", i, q(i));
-        if (q(i) < 0) {
-            logger->warn("q({}) negative", i);
-        }
-#endif
-    }
-    return q;
-}
-
-// TODO(jdstamp) ComputeHMatrix and ComputeVMatrix can probably be merged
-arma::mat ComputeHMatrix(arma::mat Sinv, std::vector<arma::mat> matrices) {
-#ifdef WITH_LOGGER_FINE
-    std::string logname = "MAPITcpp::ComputeHMatrix";
-    auto logger = spdlog::get(logname);
-    if (logger == nullptr) logger = spdlog::r_sink_mt(logname);
-    logger->info("Computing H matrix");
-#endif
-    arma::mat H = arma::zeros(arma::size(matrices[0]));
-    for (int i = 0; i < matrices.size(); i++) {
-        H = H + Sinv(0, i) * matrices[i];
-    }
-    return H;
-}
-
-arma::mat ComputeVMatrix(arma::vec delta, std::vector<arma::mat> matrices) {
-#ifdef WITH_LOGGER_FINE
-    std::string logname = "MAPITcpp::ComputeVMatrix";
-    auto logger = spdlog::get(logname);
-    if (logger == nullptr) logger = spdlog::r_sink_mt(logname);
-    logger->info("Computing V matrix");
-#endif
-    arma::mat V = arma::zeros(arma::size(matrices[0]));
-    for (int i = 0; i < matrices.size(); i++) {
-        V = V + delta(i)*matrices[i];
-    }
-    return V;
-}
 
 // TODO(jdstamp) should this be called ComputeVarianceDelta?
 double ComputeVarianceSigma(arma::vec yc, arma::mat H, arma::mat V) {
@@ -148,14 +48,6 @@ double ComputeVarianceSigma(arma::vec yc,
     }
 #endif
     return ComputeVarianceSigma(yc, H, V);
-}
-
-arma::vec RemoveFirstElement(arma::vec vector) {
-    arma::vec new_vector = arma::vec(vector.size() - 1);
-    for (int i = 1; i < vector.size(); i++) {
-        new_vector(i - 1) = vector(i);
-    }
-    return new_vector;
 }
 
 ////////////////////////////////////////////////////////////////////////////
@@ -291,10 +183,9 @@ Rcpp::List MAPITCpp(
         // Compute K and G covariance matrices
         auto start = steady_clock::now();
         // Create the linear kernel
-        arma::mat K = (GSM * p - arma::trans(X.row(i)) * X.row(i)) / (p - 1);
-        arma::mat G = K;
-        G.each_row() %= X.row(i);
-        G.each_col() %= arma::trans(X.row(i));
+        arma::rowvec x_k = X(arma::span(i), arma::span::all);
+        arma::mat K = ComputeKMatrix(GSM, x_k, p);
+        arma::mat G = ComputeGMatrix(K, x_k);
         auto end = steady_clock::now();
         execution_t(i, 0) = duration_cast<microseconds>(end - start).count();
