+++ src/MAPIT.h
@@ -18,8 +18,3 @@
 #include <RcppSpdlog>
 #include <spdlog/stopwatch.h>  // also support stopwatch feature
 #endif
-
-arma::mat GetLinearKernel(arma::mat X);
-arma::mat ComputePCs(arma::mat X, int top);
-arma::mat ComputeProjectionMatrix(int n, arma::mat b);
-double ProductTrace(arma::mat a, arma::mat b);
