+++ src/mapit/davies.cpp
@@ -1,0 +1,3 @@
+// Copyright 2021 Lorin Crawford.
+#include <RcppArmadillo.h>
+#include "davies.h"
