+++ src/mapit/projection.h
@@ -1,0 +1,4 @@
+// Copyright 2021 Lorin Crawford.
+#pragma once
+
+arma::mat ComputeProjectionMatrix(int n, arma::mat b);
