+++ src/mapit/util.h
@@ -1,0 +1,6 @@
+// Copyright 2021 Lorin Crawford.
+#pragma once
+
+arma::mat ComputePCs(arma::mat X, int top);
+
+arma::vec RemoveFirstElement(arma::vec vector);
